update masp with persistent volume

+++ scripts/launch-masp-indexer.sh
@@ -7,8 +7,8 @@ cd $HOME
 git clone https://github.com/anoma/namada-masp-indexer.git
 cd $HOME/namada-masp-indexer
 git fetch --all
-#git checkout master
-#git pull
+git checkout master
+git pull
 
 # Get the latest tag
 LATEST_TAG=$(git describe --tags $(git rev-list --tags --max-count=1))
@@ -20,7 +20,8 @@ git pull
 #cp -f $HOME/namada-campfire/docker/compose/docker-compose-namada-masp-indexer.yml $HOME/namada-masp-indexer/docker-compose.yml
 
 # prep vars
-#export POSTGRES_PORT="5433"
+export WIPE_DB=${WIPE_DB:-false}
+export POSTGRES_PORT="5435" # or 5432
 #export DATABASE_URL="postgres://postgres:password@postgres:$POSTGRES_PORT/masp_indexer_local"
 #export DATABASE_URL="postgres://postgres:password@0.0.0.0:$POSTGRES_PORT/masp_indexer_local"
 export TENDERMINT_URL=${TENDERMINT_URL:-"http://172.17.0.1:26657"}
@@ -42,11 +43,15 @@ source $env_file
 
 
 cd $HOME/namada-masp-indexer
-
-# tear down
-docker compose -f docker-compose.yml down --volumes
 docker stop $(docker container ls --all | grep 'namada-masp-' | awk '{print $1}')
 docker container rm --force $(docker container ls --all | grep 'namada-masp-' | awk '{print $1}')
+
+# tear down
+if [ "$WIPE_DB" = true ]; then
+    docker compose -f docker-compose.yml down --volumes
+else
+    docker compose -f docker-compose.yml down
+fi
 
 POSTGRES_CONTAINER_ID=$(docker ps --filter "name=postgres" --filter "publish=${POSTGRES_PORT}" --format "{{.ID}}")
 if [ -n "$POSTGRES_CONTAINER_ID" ]; then
@@ -59,12 +64,17 @@ else
     echo "No 'postgres' container found running on port ${POSTGRES_PORT} (GOOD)"
 fi
 
-if [ -z "${LOGS_NOFOLLOW}" ]; then
-    docker image rm --force $(docker image ls --all | grep 'namada-masp-' | awk '{print $3}')
-fi
+echo "Removing namada-masp-indexer images"
+docker image rm --force $(docker image ls --all | grep 'namada-masp-' | awk '{print $3}')
+docker image rm --force $(docker image ls --all | grep '<none>' | awk '{print $3}')
 
 # prune all volumes (db data)
-docker volume prune -fa
+if [ "$WIPE_DB" = true ]; then
+    docker volume prune -fa
+fi
+
+# Copy the docker compose file for the persistent db: namada-campfire/docker/compose/docker-compose-namada-masp-indexer.yml
+cp -f $HOME/namada-campfire/docker/compose/docker-compose-namada-masp-indexer.yml $HOME/namada-masp-indexer/docker-compose.yml
 
 # start up
 docker compose -f $HOME/namada-masp-indexer/docker-compose.yml --env-file $env_file up -d
